Repair TYPE_CHECKING blocks
Accomodates "if TYPE_CHECKING:" blocks that have
code in them but no imports.  Additionally adds support
for blocks that includ an "else:" block.  The "else:"
block is handled by another pass that works the same
as the "type checking" block, plus some additional logic
to maintain the former location of an if/else block as
it is removed one half at a time.

Fixes: #31
Fixes: #33